nuplayer: Fix de-referencing NULL mSource
InstantiateDecoderAction is executed after performing reset.
By the time performReset call is over, mSource is cleared and
instantiateDecoderAction is executed on a NULL mSource - causing
crash. At this point of time, mSource is already reset and it is
not intended to be used in future, so, we can safely return.
Check for NULL to avoid crash.

CRs-Fixed: 902831
Change-Id: If438da2b8c778c6db3640c0bc2e53d3ef68d286c
Signed-off-by: mydongistiny <jaysonedson@gmail.com>